Amikacin » Empiric » Cmin & Cmax

Amikacin » Empiric therapy » Total minimum and maximum concentration

simulate_amikacin_mc_emp_cmin_cmax(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  CREATININE,
  MIC,
  CMIN,
  CMAX,
  LOADINGDOSE,
  REGIMENS
)

Arguments

PATID
Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.

AGE
Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).

HEIGHT
Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).

WEIGHT
Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).

GENDER
Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').

MODEL
Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Saez Fernandez et al. (2019) - General ward').

CREATININE
Creatinine. Serum creatinine. Must be provided as numeric (min. 0.01, max. 15 mg/dL).

MIC
MIC. Minimum Inhibitory Concentration (MIC). Must be provided as numeric (min. 0.01, max. 1024 mg/L).

CMIN
Minimum concentration target. The PK/PD target can be provided as minimum blood plasma concentration (Cmin). For selecting an appropriate target please refer to the dosing guideline of your institution, e.g. <a href="http://med.stanford.edu/bugsanddrugs/guidebook/_jcr_content/main/panel_builder_584648957/panel_0/download_1194988017/file.res/Aminoglycoside%20Dosing%20Guide%202019-05-20.pdf" target="_new">Stanford Health Care Aminoglycoside Dosing Guide</a>. Must be provided as numeric (min. 0.1, max. 200 mg/L).

CMAX
Maximum concentration target. The PK/PD target can be provided as maximum blood plasma concentration (Cmax). For selecting an appropriate target please refer to the dosing guideline of your institution, e.g. <a href="http://med.stanford.edu/bugsanddrugs/guidebook/_jcr_content/main/panel_builder_584648957/panel_0/download_1194988017/file.res/Aminoglycoside%20Dosing%20Guide%202019-05-20.pdf" target="_new">Stanford Health Care Aminoglycoside Dosing Guide</a>. Must be provided as numeric (min. 0.1, max. 200 mg/L).

LOADINGDOSE
Loading dose. Loading dose is desired or not. Must be provided as string ('No' or 'Yes').

REGIMENS
Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' values. Use the regimen helper function to define the REGIMEN values.

Details

Drug: Amikacin

Method: Simulate concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

PK/PD target: Minimum and maximum blood plasma concentration (mg/L).

References

Saez Fernandez et al. (2019): Evaluation of renal function equations to predict amikacin clearance. In. Expert Review of Clinical Pharmacology. https://www.tandfonline.com/doi/full/10.1080/17512433.2019.1637253

Examples

if (FALSE) {
simulate_amikacin_mc_emp_cmin_cmax(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    MODEL = "Saez Fernandez et al. (2019) - General ward", 
    CREATININE = 1, MIC = 1, 
    CMIN = 6, CMAX = 20, 
    LOADINGDOSE = "No", 
    REGIMENS = list(list(
        set = "REGIMEN", 
        DOSE = 100, INTERVAL = 6, 
        TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 100, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 6, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 200, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 6, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 300, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 6, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 8, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 12, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 400, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 500, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 600, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 700, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 700, 
            INTERVAL = 36, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 800, 
            INTERVAL = 24, 
            TINF = 0.5), 
        list(set = "REGIMEN", 
            DOSE = 800, 
            INTERVAL = 36, 
            TINF = 0.5)))
}