Cefepime » Bayesian » % fT > MIC

Cefepime » Bayesian adaptive dosing » Percent time of free concentration above MIC

simulate_cefepime_bayesian_ftime_above_mic(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  MODEL,
  MIC,
  PCTABOVEMIC,
  CRCLCAP,
  HISTORY,
  REGIMENS
)

Arguments

PATID
Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.

AGE
Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).

HEIGHT
Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).

WEIGHT
Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).

GENDER
Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').

MODEL
Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Nicasio et al. (2009) - ICU' or 'Tam et al. (2003) - General ward').

MIC
MIC. Minimum Inhibitory Concentration (MIC). Must be provided as numeric (min. 0.01, max. 1024 mg/L).

PCTABOVEMIC
Percent of time target that the drug concentration is above MIC. The PK/PD target can be provided as the percent of time that the drug concentration is above the minimum inhibitory concentration (% T > MIC). Must be provided as numeric (min. 5, max. 100 %).

CRCLCAP
Capping Creatinine Clearance. Whether to use capping for creatinine clearance. Must be provided as string ('No cap', '120 ml/min', '130 ml/min', '140 ml/min' or '150 ml/min').

HISTORY
Historical Records. Must be provided as list of 3-24 'HISTCREATININE', 'HISTDOSE' or 'HISTCONCENTRATION' values.

REGIMENS
Dosing Regimens. List of dosing regimens to be used in simulating target attainment, from which the dosing regimen with the smallest absolute difference from the desired target will be automatically selected. Must be provided as list of 1-20 'REGIMEN' values. Use the regimen helper function to define the REGIMEN values.

Details

Drug: Cefepime

Method: Estimate the pharmacokinetic parameters of the patient from past concentrations and creatinine levels with Bayesian inverse modeling, then use that information to predict the steady state concentrations for multiple dosing regimens and select the optimal one, with regard to the target pharmacodynamic index.

PK/PD target: The percent of time that the free concentration is above the minimum inhibitory concentration.

References

Nicasio et al. (2009): Population Pharmacokinetics of High-Dose, Prolonged-Infusion Cefepime in Adult Critically Ill Patients with Ventilator-Associated Pneumonia. In. Antimicrobial Agents and Chemotherapy. https://pubmed.ncbi.nlm.nih.gov/19188394/

Tam et al. (2003): Pharmacokinetics and Pharmacodynamics of Cefepime in Patients with Various Degrees of Renal Function. In. Antimicrobial Agents and Chemotherapy. https://pubmed.ncbi.nlm.nih.gov/12760858/

K. Soetaert, T. Petzoldt (2010): Inverse Modelling, Sensitivity and Monte Carlo Analysis in R Using Package FME. In. Journal of Statistical Software. https://www.jstatsoft.org/article/view/v033i03

Examples

if (FALSE) {
simulate_cefepime_bayesian_ftime_above_mic(PATID = "Anonymous", 
    AGE = 65, HEIGHT = 175, 
    WEIGHT = 75, GENDER = "Male", 
    MODEL = "Tam et al. (2003) - General ward", 
    MIC = 4, PCTABOVEMIC = 50, 
    CRCLCAP = "No cap", 
    HISTORY = list(list(
        DATETIME = structure(1601870400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
        DOSE = 500, TINF = 2, 
        set = "HISTDOSE"), 
        list(DATETIME = structure(1601881200, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 600, 
            TINF = 1, 
            set = "HISTDOSE"), 
        list(DATETIME = structure(1601899200, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 1200, 
            TINF = 2, 
            set = "HISTDOSE"), 
        list(DATETIME = structure(1601942400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            DOSE = 750, 
            TINF = 1, 
            set = "HISTDOSE"), 
        list(DATETIME = structure(1601866800, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CREATININE = 0.9, 
            set = "HISTCREATININE"), 
        list(DATETIME = structure(1601906400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CREATININE = 0.7, 
            set = "HISTCREATININE"), 
        list(DATETIME = structure(1601888400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 10, 
            set = "HISTCONCENTRATION"), 
        list(DATETIME = structure(1601890200, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 8, 
            set = "HISTCONCENTRATION"), 
        list(DATETIME = structure(1601906400, class = c("POSIXct", 
        "POSIXt"), tzone = ""), 
            CONCENTRATION = 20, 
            set = "HISTCONCENTRATION")), 
    REGIMENS = list(list(
        DOSE = 750, INTERVAL = 8, 
        TINF = 1, set = "REGIMEN"), 
        list(DOSE = 750, 
            INTERVAL = 12, 
            TINF = 1, 
            set = "REGIMEN"), 
        list(DOSE = 750, 
            INTERVAL = 24, 
            TINF = 1, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 8, 
            TINF = 1, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 12, 
            TINF = 1, 
            set = "REGIMEN"), 
        list(DOSE = 1000, 
            INTERVAL = 24, 
            TINF = 1, 
            set = "REGIMEN"), 
        list(DOSE = 1250, 
            INTERVAL = 8, 
            TINF = 1.25, 
            set = "REGIMEN"), 
        list(DOSE = 1250, 
            INTERVAL = 12, 
            TINF = 1.25, 
            set = "REGIMEN"), 
        list(DOSE = 1250, 
            INTERVAL = 24, 
            TINF = 1.25, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 8, 
            TINF = 1.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 12, 
            TINF = 1.5, 
            set = "REGIMEN"), 
        list(DOSE = 1500, 
            INTERVAL = 24, 
            TINF = 1.5, 
            set = "REGIMEN"), 
        list(DOSE = 1750, 
            INTERVAL = 8, 
            TINF = 1.75, 
            set = "REGIMEN"), 
        list(DOSE = 1750, 
            INTERVAL = 12, 
            TINF = 1.75, 
            set = "REGIMEN"), 
        list(DOSE = 1750, 
            INTERVAL = 24, 
            TINF = 1.75, 
            set = "REGIMEN")))
}