Phenytoin » Optimal loading dose » Cmin

Phenytoin » Optimal loading dose simulation » Post-dose concentration after 24 hrs

simulate_phenytoin_mc_optloading_ncmin_24h(
  PATID,
  AGE,
  HEIGHT,
  WEIGHT,
  GENDER,
  PHENYTOINFORM,
  MODEL,
  ALBUMIN,
  RENALFUNCTIONSTATUS,
  VALPROICACID,
  CPOSTDOSE24H
)

Arguments

PATID
Patient Identifier. User-provided free text (such as patient id, name or alias) to identify related simulations. Must be provided as string.

AGE
Age. Age of the patient in years. Must be provided as numeric (min. 18, max. 120 year).

HEIGHT
Height. Height of the patient. Must be provided as numeric (min. 100, max. 250 cm).

WEIGHT
Weight. Actual body weight of the patient. Must be provided as numeric (min. 20, max. 500 kg).

GENDER
Sex. Patient's sex for clinical decision-making. Must be provided as string ('Male' or 'Female').

PHENYTOINFORM
Phenytoin form. Form of phenytoin to be used during simulations. Must be provided as string ('Phenytoin sodium IV' or 'Fosphenytoin sodium IV').

MODEL
Model for population of interest. Pharmacokinetic model to be used for specific patient type during simulations. Must be provided as string ('Muñoz et al. (2019) - ICU').

ALBUMIN
Albumin. Serum albumin. Must be provided as numeric (min. 0.01, max. 15 g/dL).

RENALFUNCTIONSTATUS
Renal function status. Renal function status of the patient. Must be provided as string ('Patients without renal impairment' or 'Dialysis or End-Stage Renal Disease').

VALPROICACID
Valproic acid blood level (leave empty if none). Total valproic acid concentration, leave empty or set to zero if no concurrent administration. Must be provided as numeric (min. 0, max. 500 mg/L).

CPOSTDOSE24H
Post-dose plasma concentration after 24 hrs target. The PK/PD target can be provided as post-dose blood plasma concentration after 24 hrs (Cmin). Must be provided as numeric (min. 0.1, max. 200 mg/L).

Details

Drug: Phenytoin

Method: Calculate optimal loading dose and simulate concentrations for the first 24 hours.

PK/PD target: Post-dose blood plasma concentration after 24 hrs (mg/L).

References

Muñoz-Pichuante, Daniel et al. (2019): Dosage of phenytoin in neurocritical patients using Bayesian algorithms: a pilot study. In. Drug Metabolism and Personalized Therapy. https://pubmed.ncbi.nlm.nih.gov/31981450

Examples

if (FALSE) {
structure(list("Anonymous", 
    45, 175, 75, "Male", 
    "Fosphenytoin sodium IV", 
    "Muñoz et al. (2019) - ICU", 
    3, "Dialysis or End-Stage Renal Disease", 
    NULL, 10), .Names = c("PATID", 
"AGE", "HEIGHT", "WEIGHT", 
"GENDER", "PHENYTOINFORM", 
"MODEL", "ALBUMIN", "RENALFUNCTIONSTATUS", 
NA, "CPOSTDOSE24H"))
}